+++ app/screens/StatsGroup.tsx
@@ -1,44 +1,16 @@
-import React, { useState, useEffect, useContext } from "react";
-import { Text } from "native-base";
-import { ScrollView, RefreshControl } from "react-native";
+import React from "react";
 import { LinearGradient } from "expo-linear-gradient";
 import Swiper from "react-native-swiper";
-import { AuthContext } from "../AuthContext";
+import Stats from "../components/Stats"; 
 
 const axios = require("axios").default;
 
 const StatsGroup = () => {
-  const auth = useContext(AuthContext);
-  let [user, setUser] = useState(null);
-  let [refreshing, setRefreshing] = useState(false);
-
-  async function fetchData() {
-    setRefreshing(true);
-    const result = await axios(
-      `https://productivitree.wl.r.appspot.com/api/v1/users/${auth.googleID}`
-    );
-    setUser(result.data.payload);
-    setRefreshing(false);
-  }
-
-  useEffect(() => {
-    fetchData();
-  }, []);
-
-  const onRefresh = React.useCallback(async () => {
-    await fetchData();
-  }, [refreshing]);
 
   return (
     <LinearGradient colors={["#C8F0EE", "#c8e2f1", "#A1C6F1"]} style={{ flex: 1 }}>
       <Swiper>
-        <ScrollView
-          refreshControl={
-            <RefreshControl refreshing={refreshing} onRefresh={onRefresh} />
-          }
-        >
-          <Text>Personal Stats</Text>
-        </ScrollView>
+        <Stats/>
         {/* <ScrollView
           refreshControl={
             <RefreshControl refreshing={refreshing} onRefresh={onRefresh} />
